Add flashcards loading and rendering
Also add a simple functional abstraction around data fetching. This app
will probably ever end up needing one, but implemented anyway as a proof
of concept how to abstract remote data fetching using vanilla React.
Adding a new component that needs data would be as simple as making a
new loader and using the `useRemoteData` effect in the component that
needs it.